[Refactor] #68 - Convertor to util
