+++ domain/src/main/java/com/woowahan/domain/usecase/InsertRecentViewedItemUseCase.kt
@@ -1,6 +1,6 @@
 package com.woowahan.domain.usecase
 
-import com.woowahan.domain.BanchanDateConverter
+import com.woowahan.domain.util.BanchanDateConvertUtil
 import com.woowahan.domain.model.BanchanModel
 import com.woowahan.domain.repository.RecentViewedRepository
 import kotlinx.coroutines.flow.Flow
@@ -10,7 +10,7 @@ class InsertRecentViewedItemUseCase (
     private val recentViewedRepository: RecentViewedRepository,
 ) {
     suspend operator fun invoke(banchan: BanchanModel, time: Date): Flow<Result<Boolean>> {
-        val timeStr= BanchanDateConverter.convert(time)
+        val timeStr= BanchanDateConvertUtil.convert(time)
         return recentViewedRepository.insertRecentViewedItem(banchan, timeStr)
     }
 }
