+++ domain/src/main/java/com/woowahan/domain/util/BanchanDateConvertUtil.kt
@@ -1,10 +1,10 @@
-package com.woowahan.domain
+package com.woowahan.domain.util
 
 import java.text.SimpleDateFormat
 import java.util.*
 
 @Suppress("SimpleDateFormat")
-object BanchanDateConverter {
+object BanchanDateConvertUtil {
     private const val banchanDateFormatString = "yyyy-MM-dd'T'HH:mm:ss'Z'"
     private val formatter = SimpleDateFormat(banchanDateFormatString).apply {
         timeZone = TimeZone.getTimeZone("UTC")
